Port over to GTK4
Initial work to port things over to GTK4. This just gets things building
and running, now we need to start rewriting some widgets. We also introduce
a temporary DrawingArea widget here to hold us over until we move on from
using cairo based rendering and start using snapshots.

+++ src/color.rs
@@ -12,7 +12,7 @@ pub const COLOR_RED: Color = Color(1.0, 0.0, 0.0);
 
 impl From<&Color> for gdk::RGBA {
     fn from(color: &Color) -> Self {
-        gdk::RGBA::new(color.0, color.1, color.2, 1.0)
+        gdk::RGBA::new(color.0 as f32, color.1 as f32, color.2 as f32, 1.0)
     }
 }
